// __generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Connection = {
  /** A list of edges (relational context) between connected nodes */
  edges: Array<Edge>;
  /** A list of connected nodes */
  nodes: Array<Node>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
};
// ...
export type Dog = ContentNode & DatabaseIdentifier & MenuItemLinkable & Node & NodeWithContentEditor & NodeWithFeaturedImage & NodeWithTemplate & NodeWithTitle & Previewable & UniformResourceIdentifiable & {
  __typename?: 'Dog';
  /** The content of the post. */
  content?: Maybe<Scalars['String']['output']>;
  /** Connection between the ContentNode type and the ContentType type */
  contentType?: Maybe<ContentNodeToContentTypeConnectionEdge>;
  /** The name of the Content Type the node belongs to */
  contentTypeName: Scalars['String']['output'];
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
  /** Post publishing date. */
  date?: Maybe<Scalars['String']['output']>;
  /** The publishing date set in GMT. */
  dateGmt?: Maybe<Scalars['String']['output']>;
  /** The desired slug of the post */
  desiredSlug?: Maybe<Scalars['String']['output']>;
  /** Added to the GraphQL Schema because the ACF Field Group &quot;Dog&quot; was set to Show in GraphQL. */
  dog?: Maybe<Dog_Dog>;
  /**
   * The id field matches the WP_Post-&gt;ID field.
   * @deprecated Deprecated in favor of the databaseId field
   */
  dogId: Scalars['Int']['output'];
  /** If a user has edited the node within the past 15 seconds, this will return the user that last edited. Null if the edit lock doesn&#039;t exist or is greater than 15 seconds */
  editingLockedBy?: Maybe<ContentNodeToEditLockConnectionEdge>;
  /** The RSS enclosure for the object */
  enclosure?: Maybe<Scalars['String']['output']>;
  /** Connection between the ContentNode type and the EnqueuedScript type */
  enqueuedScripts?: Maybe<ContentNodeToEnqueuedScriptConnection>;
  /** Connection between the ContentNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets?: Maybe<ContentNodeToEnqueuedStylesheetConnection>;
  /** Connection between the NodeWithFeaturedImage type and the MediaItem type */
  featuredImage?: Maybe<NodeWithFeaturedImageToMediaItemConnectionEdge>;
  /** The database identifier for the featured image node assigned to the content node */
  featuredImageDatabaseId?: Maybe<Scalars['Int']['output']>;
  /** Globally unique ID of the featured image assigned to the node */
  featuredImageId?: Maybe<Scalars['ID']['output']>;
  /** The global unique identifier for this post. This currently matches the value stored in WP_Post-&gt;guid and the guid column in the &quot;post_objects&quot; database table. */
  guid?: Maybe<Scalars['String']['output']>;
  /** The globally unique identifier of the danh-sach-cho object. */
  id: Scalars['ID']['output'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean']['output'];
  /** Whether the object is a node in the preview state */
  isPreview?: Maybe<Scalars['Boolean']['output']>;
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']['output']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean']['output'];
  /** The user that most recently edited the node */
  lastEditedBy?: Maybe<ContentNodeToEditLastConnectionEdge>;
  /** The permalink of the post */
  link?: Maybe<Scalars['String']['output']>;
  /** The local modified time for a post. If a post was recently updated the modified field will change to match the corresponding time. */
  modified?: Maybe<Scalars['String']['output']>;
  /** The GMT modified time for a post. If a post was recently updated the modified field will change to match the corresponding time in GMT. */
  modifiedGmt?: Maybe<Scalars['String']['output']>;
  /** Connection between the Dog type and the Dog type */
  preview?: Maybe<DogToPreviewConnectionEdge>;
  /** The database id of the preview node */
  previewRevisionDatabaseId?: Maybe<Scalars['Int']['output']>;
  /** Whether the object is a node in the preview state */
  previewRevisionId?: Maybe<Scalars['ID']['output']>;
  /** The Yoast SEO data of the ContentNode */
  seo?: Maybe<PostTypeSeo>;
  /** The uri slug for the post. This is equivalent to the WP_Post-&gt;post_name field and the post_name column in the database for the &quot;post_objects&quot; table. */
  slug?: Maybe<Scalars['String']['output']>;
  /** The current status of the object */
  status?: Maybe<Scalars['String']['output']>;
  /** The template assigned to the node */
  template?: Maybe<ContentTemplate>;
  /** The title of the post. This is currently just the raw title. An amendment to support rendered title needs to be made. */
  title?: Maybe<Scalars['String']['output']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']['output']>;
};
// ...
export type Dog_Dog = AcfFieldGroup & {
  __typename?: 'Dog_Dog';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  inStock?: Maybe<Scalars['Boolean']['output']>;
  policies?: Maybe<Scalars['String']['output']>;
  properties?: Maybe<Scalars['String']['output']>;
  sliders?: Maybe<Dog_Dog_Sliders>;
  video?: Maybe<Scalars['String']['output']>;
};
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Page_Sellingdog = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  whyChooseUs?: Maybe<Page_Sellingdog_WhyChooseUs>;
};

/** Field Group */
export type Page_Sellingdog_WhyChooseUs = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog_WhyChooseUs';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  reason1?: Maybe<Page_Sellingdog_WhyChooseUs_Reason1>;
  reason2?: Maybe<Page_Sellingdog_WhyChooseUs_Reason2>;
  reason3?: Maybe<Page_Sellingdog_WhyChooseUs_Reason3>;
  reason4?: Maybe<Page_Sellingdog_WhyChooseUs_Reason4>;
};

/** Field Group */
export type Page_Sellingdog_WhyChooseUs_Reason1 = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog_WhyChooseUs_Reason1';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Sellingdog_WhyChooseUs_Reason2 = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog_WhyChooseUs_Reason2';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Sellingdog_WhyChooseUs_Reason3 = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog_WhyChooseUs_Reason3';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Sellingdog_WhyChooseUs_Reason4 = AcfFieldGroup & {
  __typename?: 'Page_Sellingdog_WhyChooseUs_Reason4';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type Post = ContentNode & DatabaseIdentifier & MenuItemLinkable & Node & NodeWithAuthor & NodeWithComments & NodeWithContentEditor & NodeWithExcerpt & NodeWithFeaturedImage & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & NodeWithTrackbacks & Previewable & UniformResourceIdentifiable & {
// ...
  featuredImage?: Maybe<NodeWithFeaturedImageToMediaItemConnectionEdge>;
// ...
};
// ...
export type RootQueryToDogConnection = Connection & DogConnection & {
  __typename?: 'RootQueryToDogConnection';
  /** Edges for the RootQueryToDogConnection connection */
  edges: Array<RootQueryToDogConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Dog>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToDogConnectionPageInfo;
};

/** An edge in a connection */
export type RootQueryToDogConnectionEdge = DogConnectionEdge & Edge & {
  __typename?: 'RootQueryToDogConnectionEdge';
  /** A cursor for use in pagination */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The item at the end of the edge */
  node: Dog;
};
// ...
export type RootQueryToPageConnection = Connection & PageConnection & {
  __typename?: 'RootQueryToPageConnection';
  /** Edges for the RootQueryToPageConnection connection */
  edges: Array<RootQueryToPageConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Page>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToPageConnectionPageInfo;
};
// ...
export type Selling_Dog_FragmentFragment = { __typename?: 'Page', sellingDog?: { __typename?: 'Page_Sellingdog', description?: string | null, whyChooseUs?: { __typename?: 'Page_Sellingdog_WhyChooseUs', description?: string | null, reason1?: { __typename?: 'Page_Sellingdog_WhyChooseUs_Reason1', title?: string | null, description?: string | null } | null, reason2?: { __typename?: 'Page_Sellingdog_WhyChooseUs_Reason2', title?: string | null, description?: string | null } | null, reason3?: { __typename?: 'Page_Sellingdog_WhyChooseUs_Reason3', title?: string | null, description?: string | null } | null, reason4?: { __typename?: 'Page_Sellingdog_WhyChooseUs_Reason4', title?: string | null, description?: string | null } | null } | null } | null } & { ' $fragmentName'?: 'Selling_Dog_FragmentFragment' };
// ...
export type Get_Selling_DogQueryVariables = Exact<{
  name: Scalars['String']['input'];
}>;


export type Get_Selling_DogQuery = { __typename?: 'RootQuery', pages?: { __typename?: 'RootQueryToPageConnection', nodes: Array<(
      { __typename?: 'Page' }
      & { ' $fragmentRefs'?: { 'Fragment_Page_TypeFragment': Fragment_Page_TypeFragment;'Selling_Dog_FragmentFragment': Selling_Dog_FragmentFragment } }
    )> } | null };

export type Get_DogsQueryVariables = Exact<{ [key: string]: never; }>;


export type Get_DogsQuery = { __typename?: 'RootQuery', dogs?: { __typename?: 'RootQueryToDogConnection', edges: Array<{ __typename?: 'RootQueryToDogConnectionEdge', node: { __typename?: 'Dog', id: string, slug?: string | null, title?: string | null, dog?: { __typename?: 'Dog_Dog', inStock?: boolean | null, properties?: string | null, description?: string | null } | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: (
            { __typename?: 'MediaItem' }
            & { ' $fragmentRefs'?: { 'Fragment_Media_ItemFragment': Fragment_Media_ItemFragment } }
          ) } | null } }> } | null };
// ...
export type Fragment_Media_ItemFragment = { __typename?: 'MediaItem', altText?: string | null, sourceUrl?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } & { ' $fragmentName'?: 'Fragment_Media_ItemFragment' };

export type Fragment_Link_ItemFragment = { __typename?: 'AcfLink', title?: string | null, target?: string | null, url?: string | null } & { ' $fragmentName'?: 'Fragment_Link_ItemFragment' };

export type Fragment_Page_TypeFragment = { __typename?: 'Page', featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: (
      { __typename?: 'MediaItem' }
      & { ' $fragmentRefs'?: { 'Fragment_Media_ItemFragment': Fragment_Media_ItemFragment } }
    ) } | null, seo?: { __typename?: 'PostTypeSEO', title?: string | null, metaDesc?: string | null } | null } & { ' $fragmentName'?: 'Fragment_Page_TypeFragment' };